% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{analyze calls the BirdNet analayze function and returns the results or nothing}
\usage{
analyze(i, o, rtype = csv, ..., result = TRUE)
}
\arguments{
\item{i}{Path to input .wav files for BirdNet analysis.}

\item{o}{Output file location}

\item{...}{Other parameters to be passed to analyze.py}

\item{result}{Logical, determines if you want to import the result as a dataframe defaults True}
}
\value{
Nothing or a CSV file with the species detection results
}
\description{
This takes in a path to the results folder where the BirdNet results .csv
files are. This also assumes that you chose the .csv output for BirdNet.
It cannot work with other output formats! This function utilizes list.files
to get the file list recursive=T is default.
}
