% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdNet_analyze.R
\name{birdNet.analyze}
\alias{birdNet.analyze}
\title{Calls the BirdNet analayze function and returns the results as a dataframe}
\usage{
birdNet.analyze(i, o, rtype = "r", ..., result = TRUE, path)
}
\arguments{
\item{i}{Path to input .wav files for BirdNet analysis.}

\item{o}{Output file location}

\item{rtype}{The output file type. Defaults to r}

\item{...}{Other parameters to be passed to analyze.py}

\item{result}{Logical, determines if you want to import the result as a dataframe defaults True}

\item{path}{If specifying the location of BirdNet. If installed with this package, leave blank.}
}
\value{
Nothing or a CSV file with the species detection results
}
\description{
BirdNet.analyze is a wrapper for the BirdNet python script analyze.py. It
all the same parameters. See below:
\itemize{
\item i, Path to input file or folder. If this is a file, o needs to be a file too.
\item o, Path to output file or folder. If this is a file, i needs to be a file too.
\item lat, Recording location latitude. Set -1 to ignore.
\item lon, Recording location longitude. Set -1 to ignore.
\item week, Week of the year when the recording was made. Values in [1, 48] (4 weeks per month). Set -1 for year-round species list.
\item slist, Path to species list file or folder. If folder is provided, species list needs to be named "species_list.txt". If lat and lon are provided, this list will be ignored.
\item sensitivity, Detection sensitivity; Higher values result in higher sensitivity. Values in [0.5, 1.5]. Defaults to 1.0.
\item min_conf, Minimum confidence threshold. Values in [0.01, 0.99]. Defaults to 0.1.
\item overlap, Overlap of prediction segments. Values in [0.0, 2.9]. Defaults to 0.0.
\item rtype, Specifies output format. Values in ['table', 'audacity', 'r', 'csv']. Defaults to 'table' (Raven selection table).
\item threads, Number of CPU threads.
\item batchsize, Number of samples to process at the same time. Defaults to 1.
\item locale, Locale for translated species common names. Values in ['af', 'de', 'it', ...] Defaults to 'en'.
\item sf_thresh, Minimum species occurrence frequency threshold for location filter. Values in [0.01, 0.99]. Defaults to 0.03.
}
For more details see BirdNet-Analyzer on GitHub: https://github.com/kahst/BirdNET-Analyzer
}
\details{
If BirdNet was not installed using BirdNet.install() you must set path.
}
\examples{
\dontrun{
# Run without saving the result and default output
birdNet.analyze(i="D:/acoustic recordings", result=F)

# Run with default /bnResults output location
results <- birdNet.analyze(i="D:/acoustic recordings")

# Custom output location
results <- birdNet.analyze(i="D:/acoustic recordings", o="D:/results")}
}
