% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdNet_segment.R
\name{birdNet.segment}
\alias{birdNet.segment}
\title{Calls the BirdNet analayze function and returns the results as a dataframe}
\usage{
birdNet.segment(...)
}
\arguments{
\item{...}{Nothing yet}
}
\description{
BirdNet.analyze is a wrapper for the BirdNet python script analyze.py. It
all the same parameters. See below:
\itemize{
\item i, Path to input file or folder. If this is a file, o needs to be a file too.
\item audio', default='example/', help='Path to folder containing audio files.'
\item results', default='example/', help='Path to folder containing result files.'
\item o', default='example/', help='Output folder path for extracted segments.'
\item min_conf', type=float, default=0.1, help='Minimum confidence threshold. Values in [0.01, 0.99]. Defaults to 0.1.'
\item max_segments', type=int, default=100, help='Number of randomly extracted segments per species.'
\item seg_length', type=float, default=3.0, help='Length of extracted segments in seconds. Defaults to 3.0.'
\item threads', type=int, default=4, help='Number of CPU threads.'
}
For more details see BirdNet-Analyzer on GitHub: https://github.com/kahst/BirdNET-Analyzer
}
\details{
If BirdNet was not installed using BirdNet.install() you must set path.
}
