% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combRes.R
\name{combRes}
\alias{combRes}
\title{Combines BirdNet results files and adds a column for the .wave file name}
\usage{
combRes(path, rtype = "r", recursive = T)
}
\arguments{
\item{path}{The path to the folder containing the .csv files.}

\item{rtype}{The BirdNet output file type. Defaults to r.}

\item{recursive}{Defaults to True}
}
\value{
A dataframe of all .csv files in path
}
\description{
This takes in a path to the results folder where the BirdNet results .csv
files are. This also assumes that you chose the .csv output for BirdNet.
It cannot work with other output formats! This function utilizes list.files
to get the file list recursive=T is default.
}
