% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoVAD_genSynth.R
\name{ecoVAD.genSynth}
\alias{ecoVAD.genSynth}
\title{Gets the latest version of BirdNet from github}
\usage{
ecoVAD.genSynth(audioPath, speechPath, noisePath, synthPath)
}
\arguments{
\item{audioPath}{Path to the ecoacoustic training data.}

\item{speechPath}{Path to the human voices training data}

\item{noisePath}{Path to the noise training date}

\item{synthPath}{Path for the synthetic dataset}
}
\description{
getBN downloads the latest version of BirdNet from github and installs
it and the necessary python libraries in the users home directory,
unless otherwise specified. Directory ~/BirdNet-Analyzer/BirdNet-Analyzer-main.
When specifying a home directory, this same directory must be used as path in
analyze().
}
\examples{
\dontrun{
getBn()
getBn(path="C:/projectFolder/")}
}
