% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoVAD_setup.R
\name{ecoVAD.setup}
\alias{ecoVAD.setup}
\title{Creates or sets up a python virtual environment for running ecoVAD functions}
\usage{
ecoVAD.setup()
}
\description{
Creates a python virtual environment in the package directory and installs the
required python dependencies for ecoVAD to function.
}
\examples{
\dontrun{
ecoVAD.setup()}
}
