% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_python_info}
\alias{get_python_info}
\title{Get Python Information}
\usage{
get_python_info()
}
\value{
A list with the values for \code{py_path}, \code{py_cmd}, and \code{venv_activate_cmd}.
}
\description{
This function returns information about the version of Python being used by R.
}
