% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBN.R
\name{getBN}
\alias{getBN}
\title{gets latest version of BirdNet from github}
\usage{
getBN(path = path.expand("~"), remove = TRUE)
}
\arguments{
\item{path}{Path to save BirdNet, defaults to home directory.}

\item{remove}{Logical, removes or keeps .zip file, default=True}

\item{...}{Other parameters to be passed to download.file()}
}
\description{
getBN downloads the latest version of BirdNet from github and installs
it in the users home directory, unless otherwise specified.
}
