% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameRes.R
\name{renameRes}
\alias{renameRes}
\title{Renames files from different recorders to match a certain standard.}
\usage{
renameRes(
  path,
  copy_path = NULL,
  recursive = T,
  input = "auto",
  output,
  utc_offset = NULL,
  recordingName = NULL
)
}
\arguments{
\item{path}{The path to the folder containing the acoustic files you want to rename.}

\item{copy_path}{If present, path will be the input path and the files will be copied to the copy_path location with the new filename.}

\item{recursive}{Defaults to True}

\item{input}{The input file scheme, i.e. the file scheme you want to change. If auto, renameRes will attempt to detect the scheme.}

\item{output}{The output file scheme. Any value in BAR, BAR_LT, Audiomoth, SM4}

\item{utc_offset}{The UTC offset if going from any scheme to BAR_LT}

\item{recordingName}{The name, or prefix if going from Audiomoth to any other file scheme.}
}
\description{
Different ARU manufacturers have different standards for file naming
This script is renames to the files from different recorders to match
the nameing scheme of one of the other manufacturers, or a custom naming scheme
set by the user. \emph{Caution} this assumes all the files in the path are of the same
naming pattern/scheme. All files will be renamed!
}
\examples{
\dontrun{
renameRes(path="path/to/files", output=BAR_LT, utc_offset="0200", recordingName="site5")}
}
