% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vari.R
\name{vari}
\alias{vari}
\title{Outputs the frequency of each species in a BirdNet results table.}
\usage{
vari(df, overlap = 0, site = F, length)
}
\arguments{
\item{df}{A dataframe in the format of combine_results().}

\item{overlap}{The value of overlap used in the BirdNet analysis}

\item{site}{Logical, if a site column is included or not defaults False}

\item{length}{Total number of minutes of recording per day}
}
\value{
A dataframe of all .csv files in path
}
\description{
This function takes a dataframe in the format of the output of combine_results.R
with additional columns for time and date. It returns a dataframe of vocal
activity rates (VAR). is calculated by dividing the number of vocalizations
for each species by the total duration of the recording in minutes or the
number of detections per minute of recording. Date is assumed to be of class
date in the iso ISO 8601 format of YYYY-mm-dd. Time is assumed to be in the format HHMMSS. Date is
assumed to be If no overlap is provided, the function assumes that
overlap is 0.
}
