% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combRes.R
\name{combRes}
\alias{combRes}
\title{Combines BirdNet results files}
\usage{
combRes(path, rtype = "r", recursive = T)
}
\arguments{
\item{path}{The path to the folder containing the .csv files.}

\item{rtype}{The BirdNet output file type. Defaults to r.}

\item{recursive}{Defaults to True}

\item{outPath}{If included the combined results will be saved to file. If not, only returned as a dataframe.}
}
\value{
A dataframe of all files in path
}
\description{
BirdNet creates a results file for each audio file it analyzes. This takes in
a path to the results folder where the BirdNet results files are located and
combines them in to one dataframe for working with in R or saving for analysis
in other software. This function utilizes list.files to get the file list
recursive=T is default.
}
\examples{
\dontrun{
results <- combRes(path="C:/results folder/", rtype="table")

results <- combRes(path="C:/results folder/", outPath="C:/results/final/combined.csv", rtype="table")}
}
