% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBN.R
\name{getBN}
\alias{getBN}
\title{Gets the latest version of BirdNet from github}
\usage{
getBN(path, remove = TRUE, ...)
}
\arguments{
\item{path}{Path to save BirdNet, defaults to home directory.}

\item{remove}{Logical, removes the downloaded .zip file, default=True}

\item{...}{Other parameters to be passed to download.file()}
}
\description{
getBN downloads the latest version of BirdNet from github and installs
it and the necessary python libraries in the users home directory,
unless otherwise specified. Directory ~/BirdNet-Analyzer/BirdNet-Analyzer-main.
When specifying a home directory, this same directory must be used as path in
analyze().
}
\examples{
\dontrun{
getBn()
getBn(path="C:/projectFolder/")}
}
